Take a break from here
So, we have (broadly) working code for fetching definitions and translations. However,
we have a problem with pages like this one (https://cs.wiktionary.org/wiki/a), because it
has:
	(a) multiple definition (význam) sections, making it hard to reason about which
	    definition goes with which translation.
	(b) multiple languages, which makes it hard to know which things are Czech and
	    which are not.
	(c) An extra weird bug where some definitions that don't have examples aren't in
	    the results. We didn't look at this in detail since it only seemed to affect
	    non-Czech definitions in this example.

I believe that we should probably pause, take stock, and probably rewrite this stuff
using a proper HTML library that will parse it into a tree that we can handle more
easily and flexibly.